+++ FERRETERIA  Joel/Models/Producto.cs
@@ -3,7 +3,7 @@
     public class Producto
     {
         public int IdProducto { get; set; }
-        public short IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
         public string Codigo { get; set; } = string.Empty;
         public string Nombre { get; set; } = string.Empty;
         public string? Descripcion { get; set; }
@@ -14,7 +14,7 @@
         public byte Estado { get; set; }
         public DateTime FechaRegistro { get; set; }
         public DateTime? FechaActualizacion { get; set; }
-        public short IdEmpleadoResponsable { get; set; }
+        public int IdEmpleadoResponsable { get; set; }
 
         public string? NombreCategoria { get; set; }
     }
